refactor: Extract `table_row` & `tests` modules to separate files
